Change eval environment to include default packages
Enviornments in R function like a stack. your current environment has
everything you have written to it inside of it as well as a reference to
a parent envioronment which itself has a parent etc. until you get to
the base enviornment.

Originally I had used an eviornment with a parent of `baseenv()` as the
environment to run all user code, but it turns out that much of the
useful functionality is actually included in layers that are
automatically included *above* baseenv but *below* the default
enviornment. The fix is to create a new enviornment with a parent that
is the *same parent* as the default environment so that all the useful
stuff can still be in scope.